Fix for UDP received data buffering
It is not allowed to merge datagrams, implementation fixed.
If datagram doesn't fit to provided buffer then remaining part of
the datagram is lost.

Added UDP test environment and copied tests from:
 -mbed-net-socket-abstract/test
 -mbed-net-lwip

Purpose is that NS SAL tests will be merged to abstract socket tests
when NS SAL tests are mature enough.

Change-Id: I31b2e81ac52a80ce59be42264e1f385d9ac69ff3